+++ SoundWave/ConnectionStyle/SoundBoxFrequency.cs
@@ -3,16 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using raminrahimzada;
 
 namespace SoundWave.ConnectionStyle
 {
     public class SoundBoxFrequency : IOutputable
     {
-        protected readonly decimal fFrequency;
-        protected readonly decimal fFrequencyFMSpan;
+        protected readonly double fFrequency;
+        protected readonly double fFrequencyFMSpan;
         protected readonly List<IOutputable> fFrequencyFMOutputListeners = new List<IOutputable>();
-        public SoundBoxFrequency(decimal aFrequency, decimal aFrequencyFMSpan = 0)
+        public SoundBoxFrequency(double aFrequency, double aFrequencyFMSpan = 0)
         {
             fFrequency = aFrequency;
             fFrequencyFMSpan = aFrequencyFMSpan;
@@ -21,17 +20,17 @@ namespace SoundWave.ConnectionStyle
         {
             fFrequencyFMOutputListeners.Add(aFrequencyOutputable);
         }
-        public Output CalcOutput(decimal aMomentInSeconds, decimal aSampleStepDuration)
+        public Output CalcOutput(double aMomentInSeconds, double aSampleStepDuration)
         {
             Output valueFMFrequency = new Output(null, null);
             foreach (var frequencyFMOutput in fFrequencyFMOutputListeners)
             {
                 valueFMFrequency += frequencyFMOutput.CalcOutput(aMomentInSeconds, aSampleStepDuration);
             }
-            decimal timeSpan = 0m;
-            if ((valueFMFrequency.Value ?? 0m) != 0m && fFrequencyFMSpan != 0m)
+            double timeSpan = 0;
+            if ((valueFMFrequency.Value ?? 0) != 0 && fFrequencyFMSpan != 0)
             {
-                timeSpan = aSampleStepDuration * valueFMFrequency.Value * fFrequencyFMSpan ?? 0m;
+                timeSpan = aSampleStepDuration * valueFMFrequency.Value * fFrequencyFMSpan ?? 0;
             }
             return new Output(fFrequency + (fFrequencyFMSpan * valueFMFrequency.Value), Output.ConvertFrequencyToPhasePosition(aMomentInSeconds, fFrequency));
         }
